+++ src/NReco.PivotData/AverageAggregator.cs
@@ -24,7 +24,7 @@ namespace NReco.PivotData {
 	public class AverageAggregator : IAggregator {
 
 		decimal total = 0;
-		uint count = 0;
+		ulong count = 0;
 		string field;
 
 		public AverageAggregator(string f) {
@@ -35,7 +35,7 @@ namespace NReco.PivotData {
 			var stateArr = state as object[];
 			if (stateArr==null || stateArr.Length!=2)
 				throw new InvalidOperationException("Invalid state, expected array [uint count, decimal totalSum]");
-			count = Convert.ToUInt32(stateArr[0]);
+			count = Convert.ToUInt64(stateArr[0]);
 			if (stateArr[1]!=null) {
 				total = Convert.ToDecimal(stateArr[1]);
 			} else {
@@ -59,7 +59,7 @@ namespace NReco.PivotData {
 			}
 		}
 
-		public uint Count {
+		public ulong Count {
 			get { return count; }
 		}
